moved app to main-page
