+++ src/main-page/index.js
@@ -1,5 +1,5 @@
 import logo from './logo.svg';
-import './App.css';
+import './main-page.css';
 
 function App() {
   return (
